+++ models/User.js
@@ -14,10 +14,6 @@ const UserSchema = new Schema({
         type: String,
         required: true
     }
-    // date:{
-    //     type: Date,
-    //     default: Date.now
-    // }
 }, {
     timestamps: true
 })
